move travel controller interface
